AG-15850 Move `class PickManager` into a separate file
