+++ packages/ag-charts-community/src/chart/series/seriesAreaManager.ts
@@ -1,6 +1,6 @@
 import type { Point } from 'ag-charts-core';
-import { ChartUpdateType, Logger, Vec4, clamp, createId, objectsEqual } from 'ag-charts-core';
-import type { AgChartClickEvent, AgChartDoubleClickEvent, AgPickedItemsState } from 'ag-charts-types';
+import { ChartUpdateType, Logger, Vec4, clamp, createId } from 'ag-charts-core';
+import type { AgChartClickEvent, AgChartDoubleClickEvent } from 'ag-charts-types';
 
 import type {
     ActiveDatumChangeEvent,
@@ -46,6 +46,7 @@ import {
     tooltipContentAriaLabel,
 } from '../tooltip/tooltip';
 import type { UpdateOpts } from '../updateService';
+import { type IPickManager, PickManager, type PickedNode, type PickedNodes } from './pickManager';
 import { type PickFocusInputs, type PickFocusOutputs, type SeriesNodePickIntent, type UnknownSeries } from './series';
 import type { DatumIndexType, SeriesNodeDatum } from './seriesTypes';
 
@@ -76,50 +77,6 @@ export interface SeriesAreaChartDependencies {
     highlight: ChartHighlight;
     overlays: ChartOverlays;
     mode: ChartMode;
-}
-
-type PickedNodes = {
-    matches: PickedNode[];
-    distance: number;
-};
-
-interface PickedNode {
-    series: UnknownSeries;
-    datum: SeriesNodeDatum<DatumIndexType>;
-    datumIndex: unknown;
-}
-
-function getItemId(node: PickedNode): NonNullable<AgPickedItemsState['itemId']> {
-    // FIXME: How to serialise/deserialise datums is still TBD.
-    if (node.datum.itemId) return `${node.datum.itemId}`;
-    return JSON.stringify(node.datum.datumIndex);
-}
-
-type TooltipCandidate = { active?: PickedNode; paginationState?: { index: number; length: number } };
-
-/**
- * IPickManager mediates the `active` node state between SeriesAreaManager and ActiveManager.
- *
- * It tracks the active node by:
- *
- *   1.  Listening for `setState` calls from ActiveManager
- *
- *   2.  Notifying ActiveManager of state change (so that `getState` receives the
- *       correct data)
- *
- *   3.  Track tooltip candidates (if pagination is enabled).
- */
-interface IPickManager {
-    onPickedNodesHighlight(pickedNodes: PickedNodes | undefined): PickedNode | undefined;
-    onPickedNodesTooltip(pickedNodes: PickedNodes | undefined): TooltipCandidate;
-    onPickedNodesFocus(pickedFocus: PickFocusOutputs | undefined): void;
-    onPickedNodesAPI(pickedNodes: PickedNodes): void;
-    onPickedNodesAPIDebounced(): TooltipCandidate;
-
-    onClearUI(): void;
-    onClearAPI(event: null): void;
-
-    nextCandidate(): TooltipCandidate;
 }
 
 export class SeriesAreaManager extends BaseManager {
@@ -166,7 +123,7 @@ export class SeriesAreaManager extends BaseManager {
      */
     private hoverDevice: HoverDevice = 'pointer';
 
-    private readonly pickManager: IPickManager = this.initPickManager();
+    private readonly pickManager: IPickManager;
 
     private readonly focus = {
         sortedSeries: [] as UnknownSeries[],
@@ -182,6 +139,8 @@ export class SeriesAreaManager extends BaseManager {
 
     public constructor(private readonly chart: SeriesAreaChartDependencies) {
         super();
+
+        this.pickManager = new PickManager(chart.ctx.activeManager, chart.tooltip, this.focus);
 
         const initialAltText = chart.ctx.localeManager.t('ariaInitSeriesArea');
         const label1 = chart.ctx.domManager.addChild('series-area', 'series-area-aria-label1');
@@ -1178,131 +1137,6 @@ export class SeriesAreaManager extends BaseManager {
             series: desiredSeries,
         };
         return { matches: [restoredPick], distance: 0 };
-    }
-
-    private initPickManager(): IPickManager {
-        const that: SeriesAreaManager = this;
-
-        function pickedNodesEqual(a: PickedNode, b: PickedNode) {
-            return a.series === b.series && objectsEqual(a.datumIndex, b.datumIndex);
-        }
-
-        function indexOf(candidates: PickedNode[], node: PickedNode | undefined): number {
-            return node == undefined ? -1 : candidates.findIndex((c) => pickedNodesEqual(c, node));
-        }
-
-        class PickManager implements IPickManager {
-            private candidates: PickedNode[] = [];
-
-            private active: PickedNode | undefined;
-            private pendingPickedNodes?: PickedNodes;
-
-            private clear(): void {
-                this.active = undefined;
-                this.candidates.length = 0;
-                this.pendingPickedNodes = undefined;
-            }
-
-            private updateActive(active: PickedNode | undefined): PickedNode | undefined {
-                this.active = active;
-                if (this.active === undefined) {
-                    that.chart.ctx.activeManager.update({ type: 'inactive' });
-                } else {
-                    const seriesId: string = this.active.series.id;
-                    const itemId: string = getItemId(this.active);
-                    that.chart.ctx.activeManager.update({ type: 'datum', seriesId, itemId });
-                }
-                return this.active;
-            }
-
-            private popPendingPickedNodes(): PickedNodes | undefined {
-                const result = this.pendingPickedNodes;
-                this.pendingPickedNodes = undefined;
-                return result;
-            }
-
-            // Some user interactive (e.g. mouseleave, blur) has cleared the active datum.
-            onClearUI(): void {
-                that.chart.ctx.activeManager.update({ type: 'inactive' });
-                this.clear();
-            }
-
-            // Active datum was cleared by ActiveManager (`setState` or legend).
-            onClearAPI(): void {
-                this.clear();
-            }
-
-            onPickedNodesHighlight(pickedNodes: PickedNodes | undefined): PickedNode | undefined {
-                if (pickedNodes !== undefined) {
-                    const previousActive = this.active;
-                    if (that.chart.tooltip.pagination && previousActive !== undefined) {
-                        const tooltipMatch = pickedNodes.matches.find((m) => pickedNodesEqual(m, previousActive));
-                        if (tooltipMatch) {
-                            return tooltipMatch;
-                        }
-                    }
-                }
-
-                return this.updateActive(pickedNodes?.matches[0]);
-            }
-
-            onPickedNodesTooltip(pickedNodes: PickedNodes | undefined): TooltipCandidate {
-                if (pickedNodes !== undefined && that.chart.tooltip.pagination) {
-                    const previous = this.active;
-                    const nextCandidates = pickedNodes.matches;
-
-                    this.candidates = nextCandidates;
-
-                    let nextIndex = indexOf(nextCandidates, previous);
-                    if (nextIndex === -1) nextIndex = 0;
-                    this.updateActive(nextCandidates[nextIndex]);
-
-                    const paginationState = { index: nextIndex, length: nextCandidates.length };
-                    return { active: this.active, paginationState };
-                }
-
-                return { active: this.updateActive(pickedNodes?.matches[0]) };
-            }
-
-            onPickedNodesFocus(pickedFocus: PickFocusOutputs | undefined): void {
-                const { series } = that.focus;
-                this.clear();
-                if (series !== undefined && pickedFocus !== undefined) {
-                    const { datum, datumIndex } = pickedFocus;
-                    this.updateActive({ series, datum, datumIndex });
-                }
-            }
-
-            onPickedNodesAPI(debouncedPickedNodes: PickedNodes): void {
-                this.pendingPickedNodes = debouncedPickedNodes;
-            }
-
-            onPickedNodesAPIDebounced(): TooltipCandidate {
-                return { active: this.onPickedNodesHighlight(this.popPendingPickedNodes()) };
-            }
-
-            nextCandidate(): TooltipCandidate {
-                if (that.chart.tooltip.pagination) {
-                    const { candidates, active: previous } = this;
-                    const hoverIndex =
-                        previous == null ? -1 : candidates.findIndex((c) => pickedNodesEqual(c, previous));
-                    if (hoverIndex === -1) return { active: undefined, paginationState: undefined };
-
-                    let nextIndex = hoverIndex + 1;
-                    if (nextIndex >= candidates.length) {
-                        nextIndex = 0;
-                    }
-                    const nextActive = this.updateActive(candidates[nextIndex]);
-
-                    const paginationState = { index: nextIndex, length: this.candidates.length };
-                    return { active: nextActive, paginationState };
-                }
-
-                return { active: this.active };
-            }
-        }
-
-        return new PickManager();
     }
 }
